Remove redundant AgentE2E.java and instead add a class to interpret the arguments provided to tests
